+++ packages/core/src/client.ts
@@ -2,6 +2,7 @@ import {
   checkResourcesResponseFromProtobuf,
   deleteSchemasResponseFromProtobuf,
   disablePoliciesResponseFromProtobuf,
+  enablePoliciesResponseFromProtobuf,
   getPoliciesResponseFromProtobuf,
   getSchemasResponseFromProtobuf,
   listPoliciesResponseFromProtobuf,
@@ -14,8 +15,10 @@ import {
   checkResourcesRequestToProtobuf,
   deleteSchemasRequestToProtobuf,
   disablePoliciesRequestToProtobuf,
+  enablePoliciesRequestToProtobuf,
   getPoliciesRequestToProtobuf,
   getSchemasRequestToProtobuf,
+  listPoliciesRequestToProtobuf,
   planResourcesRequestToProtobuf,
 } from "./convert/toProtobuf";
 import { ValidationFailed } from "./errors";
@@ -31,11 +34,14 @@ import type {
   DeleteSchemasResponse,
   DisablePoliciesRequest,
   DisablePoliciesResponse,
+  EnablePoliciesRequest,
+  EnablePoliciesResponse,
   GetPoliciesRequest,
   GetPoliciesResponse,
   GetSchemasRequest,
   GetSchemasResponse,
   IsAllowedRequest,
+  ListPoliciesRequest,
   ListPoliciesResponse,
   ListSchemasResponse,
   PlanResourcesRequest,
@@ -418,6 +424,55 @@ export abstract class Client {
   }
 
   /**
+   * Enable multiple policies.
+   *
+   * @remarks
+   * Requires
+   *
+   * - the client to be configured with {@link Options.adminCredentials},
+   *
+   * - the Cerbos policy decision point server to be at least v0.26 and configured with the {@link https://docs.cerbos.dev/cerbos/latest/api/admin_api.html | admin API} enabled, and
+   *
+   * - a dynamic {@link https://docs.cerbos.dev/cerbos/latest/configuration/storage.html | storage backend}.
+   *
+   * @example
+   * ```typescript
+   * const result = await cerbos.enablePolicies({
+   *   ids: ["resource.document.v1", "resource.image.v1"],
+   * });
+   * ```
+   */
+  public async enablePolicies(
+    request: EnablePoliciesRequest
+  ): Promise<EnablePoliciesResponse> {
+    return enablePoliciesResponseFromProtobuf(
+      await this.admin("enablePolicy", enablePoliciesRequestToProtobuf(request))
+    );
+  }
+
+  /**
+   * Enable a policy.
+   *
+   * @remarks
+   * Requires
+   *
+   * - the client to be configured with {@link Options.adminCredentials},
+   *
+   * - the Cerbos policy decision point server to be at least v0.26 and configured with the {@link https://docs.cerbos.dev/cerbos/latest/api/admin_api.html | admin API} enabled, and
+   *
+   * - a dynamic {@link https://docs.cerbos.dev/cerbos/latest/configuration/storage.html | storage backend}.
+   *
+   * @example
+   * ```typescript
+   * const enabled = await cerbos.enablePolicy("resource.document.v1");
+   * ```
+   */
+  public async enablePolicy(id: string): Promise<boolean> {
+    const { enabledPolicies } = await this.enablePolicies({ ids: [id] });
+    return enabledPolicies === 1;
+  }
+
+  /**
    * Fetch multiple policies by ID.
    *
    * @remarks
@@ -560,11 +615,11 @@ export abstract class Client {
    * const { ids } = await cerbos.listPolicies();
    * ```
    */
-  public async listPolicies(): Promise<ListPoliciesResponse> {
+  public async listPolicies(
+    request: ListPoliciesRequest = {}
+  ): Promise<ListPoliciesResponse> {
     return listPoliciesResponseFromProtobuf(
-      await this.admin("listPolicies", {
-        includeDisabled: false,
-      })
+      await this.admin("listPolicies", listPoliciesRequestToProtobuf(request))
     );
   }